Process subprocess output as strings, not bytes
https://github.com/MestreLion/git-tools/issues/23

git-restore-mtime fails on python3 since output from subprocesses is
interpreted as bytes instead of strings.

Modify calls to subprocess methods to pass "universal_newlines=True", so that
output is returned as strings instead, which more closely resembles how python2
interprets things.  This also allows the script to still run on python2 as the
"universal_newlines" keyword is ignored.

With this change, there's no need for the python2 vs. python3 text() method, so
its definition and usages were removed as well.

Tested on Linux: python 2.7, python 3.4, python 3.5