git-restore-mtime: skip symlinks if the OS does not support updating them
If we call os.utime on a symlink, the OS will normally dereference the
symlink and update the times on the target.

To avoid this, we pass follow_symlink=False, which only works on a
subset of python versions and operating systems. If it is not supported,
skip updating the symlink since the result is probably not desirable.