Add options to git-restore-mtime for processing the log in reverse order.
This commit adds '--oldest-time' and '-o' options git-restore-mtime, which
cause it to pass the '--reverse' option to git-whatchanged, in order to process
the history log for the current branch in reverse order. This won't work in
cases where multiple files have shared the same path over a series of renames,
but this works, but should work in many other simple cases.