Correct Arch package info
My suggest to fix the incorrect instructions noted in #76. It's hard to give a one liner because the official method is `clone, build, then install`. Helper tools exist but are not universal and there are several possible ones people may prefer. Usually if people have one they know how to use it (e.g. `paru` or `yay`).