git-restore-mtime: git whatchanged has been deprecated in Git 2.51 and is scheduled for removal: git whatchanged -> git log --raw
Current documentation from https://git-scm.com/docs/git-whatchange specifies the suggested modification:
git whatchanged has been deprecated and is scheduled for removal in a future version of Git, as it is merely git log with different default; whatchanged is not even shorter to type than log --raw.
(..)
New users are encouraged to use git-log[1] instead. The whatchanged command is essentially the same as git-log[1] but defaults to showing the raw format diff output and skipping merges.